Просмотреть docker

+++ core/interaction-api/src/main/java/ru/practicum/api/dto/eventservice/EventShortDto.java
@@ -62,5 +62,5 @@ public class EventShortDto {
     /**
      * Рейтинг мероприятия.
      */
-    private long rating;
+    private double rating;
 }
